#12-DONE-Code refactor for all application
- add new styles for all app
- refactor HomeScreen
