+++ README.md
@@ -1,14 +1,8 @@
-# Thanks!
-
-Before we get started, sorry about the delay. How I said, chaotic weeks just came up with surgery included, and special thanks to Ata and Greener team for the possibility to improve my coding skills with this code challenge, which I consider very challenging for most developers
-
-I'm sure a lot of things could be improved on this challenge, (better structure for Home Screen, Private Routes, Zustand store, hide the db.json 'cause obvius reasons, environment variables, etc) but I hope you get the point what the intention of making these stuff and if you have any questions or any suggestions, I'm totally open. Nothing is more comfortable to me than the feedback which makes me improve as a developer.
-
 ## Install the project
 
 You gonna install every dependency for the project works.
 
-1. Clone the repository: `git@github.com:JPGallegos1/greener-challenge.git`
+1. Clone the repository: `git@github.com:JPGallegos1/crypto-monorepo.git`
 2. Move to `git checkout development`
 3. Create the environment variables on `[root]/packages/app/.env.local` and `[root]/apps/next/.env.local`
 4. Install dependencies `yarn install`
