+++ apps/expo/package.json
@@ -1,7 +1,7 @@
 {
   "dependencies": {
     "app": "*",
-    "expo": "^44.0.0",
+    "expo": "44.0.0",
     "expo-splash-screen": "~0.14.1",
     "expo-status-bar": "~1.2.0",
     "http": "*",
@@ -16,11 +16,11 @@
     "react-native-web": "0.17.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9",
-    "@expo/webpack-config": "^0.17.2",
+    "@babel/core": "7.12.9",
+    "@expo/webpack-config": "0.17.2",
     "@types/react": "~17.0.21",
     "@types/react-native": "~0.64.12",
-    "metro-minify-esbuild": "^0.2.0",
+    "metro-minify-esbuild": "0.2.0",
     "typescript": "~4.3.5"
   },
   "scripts": {
