+++ apps/next/package.json
@@ -9,24 +9,24 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@expo/next-adapter": "^3.1.20",
+    "@expo/next-adapter": "3.1.20",
     "app": "*",
     "http": "*",
     "next": "12.1.0",
-    "next-shield": "^2.1.3",
-    "raf": "^3.4.1",
-    "react-native-vector-icons": "^9.2.0"
+    "next-shield": "2.1.3",
+    "raf": "3.4.1",
+    "react-native-vector-icons": "9.2.0"
   },
   "devDependencies": {
-    "@expo/webpack-config": "^0.17.2",
-    "@native-base/next-adapter": "^1.1.9",
-    "@svgr/webpack": "^6.3.1",
+    "@expo/webpack-config": "0.17.2",
+    "@native-base/next-adapter": "1.1.9",
+    "@svgr/webpack": "6.3.1",
     "@types/node": "17.0.21",
-    "babel-preset-expo": "^9.2.0",
-    "eslint-config-next": "^12.1.0",
-    "next-compose-plugins": "^2.2.1",
-    "next-fonts": "^1.5.1",
-    "next-images": "^1.8.4",
-    "next-transpile-modules": "^9.0.0"
+    "babel-preset-expo": "9.2.0",
+    "eslint-config-next": "12.1.0",
+    "next-compose-plugins": "2.2.1",
+    "next-fonts": "1.5.1",
+    "next-images": "1.8.4",
+    "next-transpile-modules": "9.0.0"
   }
 }
