+++ apps/server/package.json
@@ -8,8 +8,8 @@
     "start-server": "json-server-auth db.json -p 4000"
   },
   "dependencies": {
-    "express": "^4.18.1",
-    "json-server": "^0.17.0",
-    "json-server-auth": "^2.1.0"
+    "express": "4.18.1",
+    "json-server": "0.17.0",
+    "json-server-auth": "2.1.0"
   }
 }
