--- challenge.md
@@ -1,33 +1,0 @@
-# Greener Challenge Front-End
-
-Hey JP! This is the repo where you will be working the next few days for the code challenge. There are a few things to take into account.
-
-- There is no time limit! But let's try not to think about that 😅
-- I will be recommending some libraries but you can always choose to use whichever you want 💅
-
-The goal of the challenge is seeing your amazing abilities with frontend technologies! All the assets needed to complete the tasks are provided in the `assets` folder. If there is something extra you need or I missed please let me know 😃
-
-This challenge consists of the following tasks
-
-- Create a login page with two fields, email and password. Fields must be validated with the following validations, email must be valid email and password bust be between 3 and 8 characters.
-- Create an oboarding flow that has 3 slides with some illustrations and some random text. Every app got one nowadays. This has to be done only for the mobile app.
-- Create a home screen that shows the top 10 trending crypto coins available on the market (you can use any GraphQL/REST API).
-- Create a coin detail screen, it should include it's name, price, market liquidity and whatever you feel is relevant. It would be great if you can manage to find a solution to implement a cross platform chart component.
-
-And last but not least, it's time for you to unleash your creativity and add something awesome to the app. Some questions that could spark your creativity are:
-
-- How to perform cross-platform data fetching? (Taking into account server-side (web app/nextjs) and client-side (mobile app) rendering)
-- How should we login in a cross-platform app? (Let's assume it's done with Firebase, how would it be implemented to be cross-platform)
-- How would you implement a cross-platform http client? (That is, it worked both server-side and client-side)
-
-You don't need to implement these solutions, although if you do it will be really amazing!; what we seek to know is your way of solving problems and how you approach solutions, they do not necessarily have to be perfect, our code is not... 😅
-
-The challenge will be cross-platform. Here are some libraries you might be interested in using for this challenge.
-
-- [Solito](https://solito.dev) (It's kind of a must 😂)
-- [native-base](https://nativebase.io/)
-- [react-hook-from](https://react-hook-form.com/)
-
-For navigation you should use [Solito](https://solito.dev), I would recommend to use it's starter template that sets you up with a monorepo along with the expo and next app. If you decide to set up your own environment.
-
-Good luck and have fun!
