+++ package.json
@@ -1,16 +1,16 @@
 {
-  "name": "greener-challenge",
+  "name": "ne-crypto",
   "private": true,
   "workspaces": [
     "apps/*",
     "packages/*"
   ],
   "devDependencies": {
-    "@types/react": "^17.0.39",
-    "@types/react-native": "^0.67.2",
-    "eslint": "^8.11.0",
-    "turbo": "^1.1.6",
-    "typescript": "^4.6.2"
+    "@types/react": "17.0.39",
+    "@types/react-native": "0.67.2",
+    "eslint": "8.11.0",
+    "turbo": "1.1.6",
+    "typescript": "4.6.2"
   },
   "scripts": {
     "native": "cd apps/expo && expo start",
