+++ packages/app/components/CoinLink.tsx
@@ -6,7 +6,7 @@ export function CoinLink(props: PropsWithChildren<UseLinkProps>) {
   const linkProps = useLink(props)
 
   return (
-    <Pressable _hover={{ bg: 'gray.200' }} {...props} {...linkProps}>
+    <Pressable _hover={{ bg: '#E5E5E5' }} {...props} {...linkProps}>
       {props.children}
     </Pressable>
   )
