--- packages/app/components/Logo.tsx
@@ -1,19 +1,0 @@
-import * as React from 'react'
-import Svg, { Path } from 'react-native-svg'
-
-const Logo = (props) => (
-  <Svg
-    width={250}
-    height={80}
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <Path
-      d="M42.591 34.648a25.587 25.587 0 0 0-6.268-5.87l-.925-.642c-2.473-1.683-4.81-3.384-4.81-3.384a31.8 31.8 0 0 1-8.561-9.139A23.922 23.922 0 0 1 18.649 7a77.465 77.465 0 0 0-6.12 15.555c-3.14 12-.469 17.757-.469 17.757a22.375 22.375 0 0 0 18.682 15.24 19.356 19.356 0 0 0 8.516-.97c-.584-.135-15.978-3.853-16.91-27.834v-.212c0-.071 2.929 18.547 18.143 27.506a9.438 9.438 0 0 0 4.688-4.656c2.621-5.728.154-11.086-2.588-14.738ZM72.108 51.593a8.596 8.596 0 0 0 6.107-2.627 8.44 8.44 0 0 0 2.819-6.273v-.051a13.487 13.487 0 0 1-18.47-.592 13.02 13.02 0 0 1-3.956-9.542 13.025 13.025 0 0 1 3.956-9.544 13.487 13.487 0 0 1 19.067 0 13.017 13.017 0 0 1 3.975 9.525v10.2a12.844 12.844 0 0 1-4.251 9.542 13.017 13.017 0 0 1-9.247 3.951c-5.292 0-9.248-2.044-11.868-6.133a2.274 2.274 0 0 1 1.139-3.367h.077a2.087 2.087 0 0 1 2.646.886 8.972 8.972 0 0 0 8.006 4.026Zm8.927-19.267a8.924 8.924 0 0 0-17.847.173 8.928 8.928 0 1 0 17.847 0v-.172ZM102.125 19.002a2.192 2.192 0 0 1 1.606.642 2.238 2.238 0 0 1 .642 1.606v.05a2.16 2.16 0 0 1-.642 1.587 2.19 2.19 0 0 1-1.606.642h-1.406a5.363 5.363 0 0 0-3.956 1.63 5.483 5.483 0 0 0-1.631 4.008v14.585a2.1 2.1 0 0 1-.642 1.58 2.158 2.158 0 0 1-1.586.642h-.071a2.158 2.158 0 0 1-1.586-.641 2.099 2.099 0 0 1-.642-1.58V29.168a10.114 10.114 0 0 1 2.967-7.2 9.807 9.807 0 0 1 7.2-2.968h1.353ZM132.116 26.89a2.221 2.221 0 0 1-1.111 2.992l-9.1 4.1-9.075 4.1a8.492 8.492 0 0 0 7.026 3.417 8.926 8.926 0 0 0 8.422-6.01 2.122 2.122 0 0 1 2.171-1.534 2.175 2.175 0 0 1 1.856.97 2.13 2.13 0 0 1 .27 2.048 12.984 12.984 0 0 1-4.842 6.48 13.2 13.2 0 0 1-7.886 2.57 12.995 12.995 0 0 1-9.537-3.976 12.984 12.984 0 0 1-3.956-9.542 13.003 13.003 0 0 1 3.956-9.544 13.488 13.488 0 0 1 16.929-1.76 13.198 13.198 0 0 1 4.877 5.689Zm-12.26-3.262a8.925 8.925 0 0 0-8.574 6.44 8.936 8.936 0 0 0-.257 3.796l15.653-7.066a8.495 8.495 0 0 0-6.82-3.173l-.002.003ZM162.211 26.884a2.232 2.232 0 0 1-1.118 3l-9.094 4.1-9.075 4.1a8.473 8.473 0 0 0 7.019 3.416 8.947 8.947 0 0 0 8.432-6.01 2.122 2.122 0 0 1 2.177-1.535 2.183 2.183 0 0 1 1.856.97 2.12 2.12 0 0 1 .27 2.049 13 13 0 0 1-4.849 6.48 13.185 13.185 0 0 1-7.886 2.569 13.004 13.004 0 0 1-9.543-3.976 12.984 12.984 0 0 1-3.956-9.542 13.002 13.002 0 0 1 3.956-9.544 13.486 13.486 0 0 1 16.91-1.752 13.173 13.173 0 0 1 4.9 5.677l.001-.002Zm-12.266-3.262a8.897 8.897 0 0 0-8.927 8.953c-.001.43.032.86.1 1.284l15.651-7.065a8.519 8.519 0 0 0-6.824-3.169v-.003ZM171.851 43.752a2.07 2.07 0 0 1-.642 1.58 2.134 2.134 0 0 1-1.58.642h-.077a2.142 2.142 0 0 1-1.58-.641 2.1 2.1 0 0 1-.642-1.58V30.226a10.845 10.845 0 0 1 3.288-7.938 11.197 11.197 0 0 1 15.85 0 10.836 10.836 0 0 1 3.288 7.938v13.525A2.279 2.279 0 0 1 187.508 46h-.051a2.164 2.164 0 0 1-1.606-.642 2.077 2.077 0 0 1-.642-1.58v-13.55a6.635 6.635 0 0 0-1.916-4.749 6.642 6.642 0 0 0-11.365 4.748l-.077 13.525ZM219.67 26.888a2.224 2.224 0 0 1-1.111 2.993l-9.113 4.1-9.075 4.1a8.478 8.478 0 0 0 7.019 3.417 8.925 8.925 0 0 0 8.477-6.01 2.127 2.127 0 0 1 2.177-1.536 2.176 2.176 0 0 1 1.856.97 2.12 2.12 0 0 1 .27 2.05 13.021 13.021 0 0 1-4.842 6.48 13.199 13.199 0 0 1-7.893 2.568 12.994 12.994 0 0 1-9.55-3.975 13.008 13.008 0 0 1-3.956-9.543 13.035 13.035 0 0 1 3.956-9.55 13.486 13.486 0 0 1 16.916-1.753 13.203 13.203 0 0 1 4.869 5.69Zm-12.266-3.262a8.933 8.933 0 0 0-8.567 6.443 8.929 8.929 0 0 0-.257 3.794l15.651-7.064a8.502 8.502 0 0 0-6.827-3.173ZM236.309 19.002a2.192 2.192 0 0 1 1.606.642 2.24 2.24 0 0 1 .642 1.606v.05a2.145 2.145 0 0 1-.642 1.587 2.19 2.19 0 0 1-1.606.642h-1.407a5.366 5.366 0 0 0-3.956 1.63 5.51 5.51 0 0 0-1.638 3.957v14.636a2.085 2.085 0 0 1-.642 1.58 2.16 2.16 0 0 1-1.586.642h-.071a2.154 2.154 0 0 1-1.586-.641 2.101 2.101 0 0 1-.642-1.58V29.168a10.057 10.057 0 0 1 2.967-7.2 9.811 9.811 0 0 1 7.2-2.967h1.361Z"
-      fill="#33EB70"
-    />
-  </Svg>
-)
-
-export default Logo
