+++ packages/app/package.json
@@ -3,28 +3,28 @@
   "name": "app",
   "main": "index.ts",
   "dependencies": {
-    "@hookform/resolvers": "^2.9.7",
-    "@react-native-async-storage/async-storage": "^1.17.9",
-    "@react-navigation/bottom-tabs": "^6.3.2",
-    "@react-navigation/native": "^6.0.11",
-    "@react-navigation/native-stack": "^6.5.0",
-    "@types/d3-color": "^3.1.0",
-    "expo-linking": "^3.0.0",
-    "moti": "^0.19.0-alpha.6",
-    "native-base": "^3.4.11",
-    "react-hook-form": "^7.34.0",
-    "react-native-onboarding-swiper": "^1.2.0",
+    "@hookform/resolvers": "2.9.7",
+    "@react-native-async-storage/async-storage": "1.17.9",
+    "@react-navigation/bottom-tabs": "6.3.2",
+    "@react-navigation/native": "6.0.11",
+    "@react-navigation/native-stack": "6.5.0",
+    "@types/d3-color": "3.1.0",
+    "expo-linking": "3.0.0",
+    "moti": "0.19.0-alpha.6",
+    "native-base": "3.4.11",
+    "react-hook-form": "7.34.0",
+    "react-native-onboarding-swiper": "1.2.0",
     "react-native-svg": "12.1.1",
-    "react-native-web": "^0.18.7",
+    "react-native-web": "0.18.7",
     "solito": "latest",
-    "victory": "^36.6.0",
-    "victory-native": "^36.6.0",
-    "yup": "^0.32.11",
-    "zustand": "^4.0.0"
+    "victory": "36.6.0",
+    "victory-native": "36.6.0",
+    "yup": "0.32.11",
+    "zustand": "4.0.0"
   },
   "sideEffects": false,
   "devDependencies": {
-    "@expo/webpack-config": "^0.17.2",
-    "babel-loader": "^8.2.5"
+    "@expo/webpack-config": "0.17.2",
+    "babel-loader": "8.2.5"
   }
 }
